+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Changelog
  * remove some ansible playbook filenames
  * improve frontend calendar view
  * disable show previous editions if there are no related activities condition
+ * enable show related activities if there are no previous editions condition
 
 ##### Version 5.09.01 (2022-03-01)
  * Symfony 5.4.5 update